food-validator (ML)/requirements.txt
tensorflow==2.4.1
flask
numpy
Pillow